% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWBLS.R
\name{CWBLS}
\alias{CWBLS}
\title{Estimate true causal effect in bivariate Mendelian randomization to adjust for collider bias and weak instrument bias.}
\usage{
CWBLS(data)
}
\arguments{
\item{data}{A data frame of input dataset. This dataset should contain following columns: \code{xbeta} and \code{xse} represents the effects of exposure trait and its standard errors, \code{dbeta} and \code{dse} represents the effects of disease trait and its standard errors, \code{ybeta} and \code{yse} represents the effects of disease progression trait and its standard errors.}
}
\value{
An object of class "CWBLS" which contains:
  \itemize{
    \item{\code{b} The true causal estimate between exposure and disease progression}
    \item{\code{b.se} Standard error of \code{b}}
  }
}
\description{
This function is designed for adjusting collider bias and weak instrument bias to estimate unbiased causal effect between an exposure trait and disease progression trait,
using generalised instrumental effect regression and CWBLS adjustment in bivariate Mendelian randomization.
}
\examples{

# Load the test dataset
data(testData)

# Find the true causal between exposure and disease progression.
CWBLS(testData)
}
\references{
Cai S, Hartley A, Mahmoud O, Tilling K, Dudbridge F. Adjusting for collider bias in genetic association studies using instrumental variable methods. Genet Epidemiol. 2022;46:303–16.
}
\author{
Siyang Cai, Frank Dudbridge
}
