% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSteps.R
\name{TwoSteps}
\alias{TwoSteps}
\title{Adjust causal effect for collider bias in disease progression}
\usage{
TwoSteps(data, method, od = FALSE)
}
\arguments{
\item{data}{A data frame ofinput dataset. This dataset should contain following columns: \code{xbeta} and \code{xse} represents the effects of exposure trait and its standard errors, \code{dbeta} and \code{dse} represents the effects of disease trait and its standard errors, \code{ybeta} and \code{yse} represents the effects of disease progression trait and its standard errors}

\item{method}{Adjustment method in each step to correct for collider bias and weak instrument bias. Choose from \code{cwls} and \code{mr.raps}. If MR-RAPS is used, an argument \code{od = TRUE} is needed if invalid instruments are used.}
}
\value{
An object of class "methodCB" which contains:
  \itemize{
    \item{\code{b} The true causal estimate between exposure and disease progression}
    \item{\code{b.se} Standard error of \code{b}}
  }
}
\description{
This function is designed for adjusting collider bias and weak instrument bias to estimate unbiased causal effect between an exposure trait and disease progression trait,
using two-step regression and CWLS/MR-RAPS adjustment.
}
\examples{

# Load the test dataset
data(testData)

# Find the true causal between exposure and disease progression, using two-step regression with CWLS.
TwoSteps(testData, method = "cwls")

# Similarly, we can replace the method by MR-RAPS. But note that we are using invalid instrument so over-dispersion needs to be considered.
TwoSteps(testData, method = "mr.raps", od = TRUE)

}
\references{
Cai S, Hartley A, Mahmoud O, Tilling K, Dudbridge F. Adjusting for collider bias in genetic association studies using instrumental variable methods. Genet Epidemiol. 2022;46:303–16.

Zhao, Q, et al. "Statistical inference in two-sample summary-data Mendelian randomization using robust adjusted profile score." The Annals of Statistics 48.3 (2020): 1742-1769.
}
\author{
Siyang Cai, Frank Dudbridge
}
